# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ImpactSeverity": {
        "description": "Deterministic or AI-assisted impact rating.",
        "enum": [
          "none",
          "info",
          "warning",
          "critical"
        ],
        "title": "ImpactSeverity",
        "type": "string"
      },
...
      "ObjectVersionResponse": {
        "description": "Secret-redacted immutable object version.",
        "properties": {
          "actor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor"
          },
          "audit_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audit Id"
          },
          "changed_fields": {
            "items": {
              "type": "string"
            },
            "title": "Changed Fields",
            "type": "array"
          },
          "configuration": {
            "additionalProperties": true,
            "title": "Configuration",
            "type": "object"
          },
          "configuration_hash": {
            "title": "Configuration Hash",
            "type": "string"
          },
          "event": {
            "$ref": "#/components/schemas/VersionEvent"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_deleted": {
            "title": "Is Deleted",
            "type": "boolean"
          },
          "observed_at": {
            "format": "date-time",
            "title": "Observed At",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "version",
          "event",
          "configuration",
          "configuration_hash",
          "changed_fields",
          "is_deleted",
          "observed_at",
          "actor",
          "audit_id"
        ],
        "title": "ObjectVersionResponse",
        "type": "object"
      },
...
      "PointInTimeModeResponse": {
        "description": "Whether the caller is browsing a past point in time.",
        "properties": {
          "as_of": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of"
          },
          "historical": {
            "title": "Historical",
            "type": "boolean"
          }
        },
        "required": [
          "historical"
        ],
        "title": "PointInTimeModeResponse",
        "type": "object"
      },
      "PointInTimeStateResponse": {
        "description": "A scope reconstructed as it existed at one instant.",
        "properties": {
          "as_of": {
            "format": "date-time",
            "title": "As Of",
            "type": "string"
          },
          "object_count": {
            "title": "Object Count",
            "type": "integer"
          },
          "objects": {
            "items": {
              "$ref": "#/components/schemas/ReconstructedObjectResponse"
            },
            "title": "Objects",
            "type": "array"
          },
          "scope": {
            "enum": [
              "org",
              "site"
            ],
            "title": "Scope",
            "type": "string"
          }
        },
        "required": [
          "as_of",
          "scope",
          "object_count"
        ],
        "title": "PointInTimeStateResponse",
        "type": "object"
      },
...
      "PydanticObjectId": {
        "example": "5eb7cf5a86d9755df3a6c593",
        "maxLength": 24,
        "minLength": 24,
        "pattern": "^[0-9a-f]{24}$",
        "type": "string"
      },
...
      "ReconstructedObjectResponse": {
        "description": "One object as it existed at the reconstructed instant.",
        "properties": {
          "is_deleted": {
            "default": false,
            "title": "Is Deleted",
            "type": "boolean"
          },
          "logical_object_id": {
            "title": "Logical Object Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "object_type": {
            "title": "Object Type",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          },
          "version_id": {
            "title": "Version Id",
            "type": "string"
          }
        },
        "required": [
          "logical_object_id",
          "object_type",
          "name",
          "version_id",
          "version"
        ],
        "title": "ReconstructedObjectResponse",
        "type": "object"
      },
...
      "TimelineMarkerListResponse": {
        "description": "Every marker inside the selected window.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TimelineMarkerResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "range_end": {
            "format": "date-time",
            "title": "Range End",
            "type": "string"
          },
          "range_start": {
            "format": "date-time",
            "title": "Range Start",
            "type": "string"
          }
        },
        "required": [
          "range_start",
          "range_end"
        ],
        "title": "TimelineMarkerListResponse",
        "type": "object"
      },
      "TimelineMarkerResponse": {
        "description": "One change tick drawn on the shell's time bar.",
        "properties": {
          "at": {
            "format": "date-time",
            "title": "At",
            "type": "string"
          },
          "change_group_id": {
            "title": "Change Group Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "severity": {
            "$ref": "#/components/schemas/ImpactSeverity",
            "default": "none"
          }
        },
        "required": [
          "at",
          "change_group_id",
          "label"
        ],
        "title": "TimelineMarkerResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "VersionEvent": {
        "description": "Reason an immutable object version was created.",
        "enum": [
          "initial",
          "created",
          "updated",
          "deleted",
          "restored"
        ],
        "title": "VersionEvent",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "CsrfToken": {
        "description": "Double-submit CSRF token. Required on every unsafe method of a cookie session.",
        "in": "header",
        "name": "X-CSRF-Token",
        "type": "apiKey"
      },
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        },
        "type": "oauth2"
      },
      "SessionCookie": {
        "description": "Browser session issued by `POST /auth/login`. Unsafe methods must also send the `cg_csrf` cookie value in the `X-CSRF-Token` header.",
        "in": "cookie",
        "name": "cg_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
...
    "summary": "Configuration history, point-in-time recovery, and impact monitoring for Juniper Mist.",
    "title": "Mist Config Guardian",
    "version": "0.2.2"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/organizations/{organization_id}/point-in-time/markers": {
      "get": {
        "description": "List the change markers drawn on the shell's time bar.",
        "operationId": "list_markers_api_v1_organizations__organization_id__point_in_time_markers_get",
        "parameters": [
          {
            "in": "path",
            "name": "organization_id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PydanticObjectId"
            }
          },
          {
            "in": "query",
            "name": "range",
            "required": false,
            "schema": {
              "default": "24h",
              "enum": [
                "24h",
                "7d",
                "30d"
              ],
              "title": "Range",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Config-Guardian-As-Of",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Config-Guardian-As-Of"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineMarkerListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "CsrfToken": [],
            "SessionCookie": []
          }
        ],
        "summary": "List Markers",
        "tags": [
          "Point In Time"
        ]
      }
    },
    "/api/v1/organizations/{organization_id}/point-in-time/mode": {
      "get": {
        "description": "Report whether this request is browsing a reconstructed past instant.",
        "operationId": "read_mode_api_v1_organizations__organization_id__point_in_time_mode_get",
        "parameters": [
          {
            "in": "path",
            "name": "organization_id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PydanticObjectId"
            }
          },
          {
            "in": "header",
            "name": "X-Config-Guardian-As-Of",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Config-Guardian-As-Of"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PointInTimeModeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "CsrfToken": [],
            "SessionCookie": []
          }
        ],
        "summary": "Read Mode",
        "tags": [
          "Point In Time"
        ]
      }
    },
    "/api/v1/organizations/{organization_id}/point-in-time/objects/{logical_object_id}": {
      "get": {
        "description": "Return one object's configuration as it existed at an instant.",
        "operationId": "read_object_at_api_v1_organizations__organization_id__point_in_time_objects__logical_object_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "logical_object_id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PydanticObjectId"
            }
          },
          {
            "in": "path",
            "name": "organization_id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PydanticObjectId"
            }
          },
          {
            "in": "query",
            "name": "at",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "At"
            }
          },
          {
            "in": "header",
            "name": "X-Config-Guardian-As-Of",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Config-Guardian-As-Of"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectVersionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "CsrfToken": [],
            "SessionCookie": []
          }
        ],
        "summary": "Read Object At",
        "tags": [
          "Point In Time"
        ]
      }
    },
    "/api/v1/organizations/{organization_id}/point-in-time/state": {
      "get": {
        "description": "Reconstruct an organization or site as it existed at an instant.",
        "operationId": "read_state_api_v1_organizations__organization_id__point_in_time_state_get",
        "parameters": [
          {
            "in": "path",
            "name": "organization_id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PydanticObjectId"
            }
          },
          {
            "in": "query",
            "name": "at",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "At",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "default": "org",
              "enum": [
                "org",
                "site"
              ],
              "title": "Scope",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "site_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Site Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PointInTimeStateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "CsrfToken": [],
            "SessionCookie": []
          }
        ],
        "summary": "Read State",
        "tags": [
          "Point In Time"
        ]
      }
    },
...
  },
...
}